# le-backend/docs/openapi_enhanced.yaml
openapi: 3.0.3
info:
  title: LC Workflow API - Enhanced
  description: |
    Enhanced LC Workflow API with comprehensive system stability improvements, 
    security features, monitoring capabilities, and data synchronization.
    
    ## Features
    
    - **Enhanced File Management**: Secure file upload with malware scanning and encryption
    - **Folder Organization**: Automatic document organization by type
    - **Health Monitoring**: Comprehensive system health checks and metrics
    - **Data Synchronization**: Real-time updates and cache management
    - **Security**: Advanced authentication, authorization, and audit logging
    - **Feature Flags**: Gradual rollout and A/B testing capabilities
    
    ## Authentication
    
    All endpoints require JWT authentication unless specified otherwise.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Error Handling
    
    The API uses standardized error responses with correlation IDs for tracking.
    All errors include suggested actions and detailed information.
    
    ## Rate Limiting
    
    - File Upload: 100 requests/hour per user
    - General API: 1000 requests/hour per user
    - Health Checks: 60 requests/minute (no user limit)
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /monitoring/alerts/active:
    get:
      tags:
        - Monitoring
      summary: Get active alerts
      description: Retrieve active system alerts
      operationId: getActiveAlerts
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, warning, info]
          description: Filter by severity
      responses:
        '200':
          description: Active alerts retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveAlertsResponse'
  /monitoring/alerts/{alert_id}/acknowledge:
    post:
      tags:
        - Monitoring
      summary: Acknowledge alert
      description: Acknowledge an active alert
      operationId: acknowledgeAlert
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  acknowledged_by:
                    type: string
                  timestamp:
                    type: string
                    format: date-time

  # Data Synchronization Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        component:
          type: string
        timestamp:
          type: string
          format: date-time
        acknowledgments:
          type: array
          items:
            type: string
        details:
          type: object
    ActiveAlertsResponse:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        total_active:
          type: integer
        severity_filter:
          type: string
          nullable: true

    # Data Synchronization Schemas
...